% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{calc.beta.oracle}
\alias{calc.beta.oracle}
\title{Calculate systematic effects model with full potential outcomes}
\usage{
calc.beta.oracle(Y1, Y0, Z, formula, data, method = c("RI", "OLS"))
}
\arguments{
\item{Y1}{Full list of Y(1)}

\item{Y0}{Full list of Y(0)}

\item{Z}{Realized treatment assignment}

\item{formula}{Formula for the covariates X}

\item{data}{Dataframe with covariates}

\item{method}{RI or OLS.}
}
\description{
Calculate the actual beta for our sample using the full
potential outcome schedule (if available)
}
\details{
This function is useful for sanity checks and simulation studies
to compare real estimators to optimal baselines.

Not useful for observed data as it relys on complete knowledge of all potential outcomes.
}
\seealso{
make.randomized.dat
covariance between X and tau
covariance between X and Y(0)
covariance between X and Y(1)
Calculate individual systematic effects
return results
}
