% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{coef.RI.regression.result}
\alias{coef.RI.regression.result}
\title{Extract coefficients of a fit RI regression model.}
\usage{
\method{coef}{RI.regression.result}(x)
}
\arguments{
\item{x}{A RI.regression.result object.}
}
\description{
Extract coefficients of a fit RI regression model.
}
