% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{est.beta.LATE}
\alias{est.beta.LATE}
\title{Estimate systematic effects in presence of noncompliance}
\usage{
est.beta.LATE(formula, data, interaction.formula, method = c("RI",
  "2SLS"), na.rm = TRUE)
}
\arguments{
\item{formula}{An object of class formula, as in lm(), such as Y ~ D | Z with only the endogenous variable (D) and the instrument (Z) on the right-hand side separated by a vertical bar (|).}

\item{data}{A data.frame, tbl_df, or data.table with the input data.}

\item{interaction.formula}{A right-sided formula with pre-treatment covariates to model treatment effects for on the right hand side, such as ~ x1 + x2 + x3.}

\item{method}{RI or 2SLS.}

\item{na.rm}{A logical flag indicating whether to list-wise delete missing data. The function will report an error if missing data exist. Default is FALSE.}
}
\description{
Estimate a model of effects for compliers under exclusion restriction for
never- and always-takers and under monotonicity.
}
\seealso{
est.beta
sample size, and sub-sample sizes classified by (Z, D)
estimation of the proportions of the latent strata
TSLS
residual
gamma for compliers: under treatment
gamma for compliers: under control
unbiased estimator for beta
residual
covariance of beta
testing
return
}
