% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{est.beta}
\alias{est.beta}
\title{Estimate Systematic Treatment Effect Variation}
\usage{
est.beta(formula, data, interaction.formula, control.formula = NULL,
  method = c("RI", "OLS"), empirical.Sxx = FALSE, na.rm = FALSE)
}
\arguments{
\item{formula}{An object of class formula, as in lm(), such as Y ~ Z with only the treatment variable on the right-hand side.}

\item{data}{A data.frame, tbl_df, or data.table with the input data.}

\item{interaction.formula}{A right-sided formula with pre-treatment covariates to model treatment effects for on the right hand side, such as ~ x1 + x2 + x3.}

\item{control.formula}{A right-sided formula with pre-treatment covariates to adjust for on the right hand side, such as ~ x1 + x2 + x3. Default is NULL (no variables adjusted for). Default is NULL}

\item{method}{RI or OLS.  method=OLS is shorthand for setting the empirical.Sxx variable to TRUE, nothing more.}

\item{empirical.Sxx}{Estimate seperate Sxx for treatment and control if TRUE, use known Sxx if not.}

\item{na.rm}{A logical flag indicating whether to list-wise delete missing data. The function will report an error if missing data exist. Default is FALSE.}
}
\description{
Estimate our treatment effect model given by the formula "~ X"
Will do the different estimators listed in our paper given the 'method' flag
}
\seealso{
print.RI.regression.result
sample size
dimension of X
treatment group
control group
calculate covariance matrix of X and estimate our gammas
residuals
estimate beta
final estimator for tau, the individual systematic treatments
covariance
testing whether any of the non-intercept terms are nonzero
return results
}
