% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wrappers.R
\name{fishpidetect}
\alias{fishpidetect}
\title{fishpidetect}
\usage{
fishpidetect(Y, Z, W, X, plugin, tau.hat, test.stat, te.vec,
B, gamma, grid.gamma, grid.size, return.matrix, n.cores, verbose, ...)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}

\item{W}{Additional pre-treatment covariates to interact with T to define linear model of treatment effects. Default is NULL.}

\item{X}{Additional pre-treatment covariates to adjust for in estimation, but not to interact with treatment. Default is NULL.}

\item{plugin}{Whether to calculate the plug-in p-value without sweeping over range of possible treatment effect magnitudes. Default is FALSE.}

\item{tau.hat}{The value of the plug-in treatment effect. Default is sample average treatment effect.}

\item{test.stat}{Test statistic function to use on the data. Default is shifted Kolmogorov-Smirnov statistic.}

\item{te.vec}{Vector of taus to examine if you want to override generating ones automatically. Default is NULL.}

\item{B}{Number of permutations to take. Default is 500.}

\item{gamma}{How wide of a CI to make around tau-hat for search. Default is 0.0001.}

\item{grid.gamma}{Parameter to govern where the grid points are sampled.  Bigger values means more samples towards the estimated tau-hat. Default is 100*gamma.}

\item{grid.size}{Number of points in the grid. Default is 151.}

\item{return.matrix}{Whether to return the matrix of all the imputed statistics.  Default is FALSE.}

\item{n.cores}{Number of cores to use to parallelize permutation step. Default is 1.}

\item{verbose}{Whether to print out progress bar when fitting and other diagnostics. Default is TRUE.}

\item{...}{Extra arguments passed to the generate.permutations function and test.stat functions.}
}
\value{
If plug-in, the value of the test and the associated p-value. If not, a list with the value of
the test statistic on the observed data, the value of the CI-adjusted p-value, the plug-in p-value, and other information on the test.
}
\description{
Test for systematic treatment effect heterogeneity using
Fisherian permutation inference methods.
}
\examples{
Z <- rep(c(0, 1), 100)
tau <- 4
Y <- ifelse(Z, rnorm(100, tau), rnorm(100, 0))
tst <- fishpidetect(Y, Z, B = 50, grid.size = 50)
}
