% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wrappers.R
\name{FRTCI.interact}
\alias{FRTCI.interact}
\title{FRTCI.interact}
\usage{
FRTCI.interact(Y, Z, W, X, test.stat, B, gamma, grid.gamma, grid.size, return.matrix,
n.cores, verbose, ...)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector (1=Tx, 0=Co)}

\item{W}{Additional pre-treatment covariates to interact with T to increase precision of treatment effect estimate.}

\item{X}{Additional pre-treatment covariates to adjust for in estimation, but not to interact with treatment.}

\item{test.stat}{Test statistic function to use on the data. Default is shifted Kolmogorov-Smirnov statistic.}

\item{B}{Number of permutations to take. Default is 500.}

\item{gamma}{How wide of a CI to make around tau-hat for search. Default is 0.0001.}

\item{grid.gamma}{Parameter to govern where the grid points are sampled.  Bigger values means more samples towards the estimated tau-hat. Default is 100*gamma.}

\item{grid.size}{Number of points in the grid. Default is 151.}

\item{return.matrix}{Whether to return the matrix of all the imputed statistics.  Default is FALSE.}

\item{n.cores}{Number of cores to use to parallelize permutation step. Default is 1.}

\item{verbose}{Whether to print out progress bar when fitting and other diagnostics. Default is TRUE.}

\item{...}{Extra arguments passed to the generate.permutations function (and the test.stat function).}
}
\value{
A list with the value of the test statistic on the observed data, the value of the CI-adjusted p-value,
the plug-in p-value, and other information on the test.
}
\description{
Conduct the FRT CI Method, adjusting for covariates using a linear model, on passed data
for a model of effects W'beta with unknown beta.
}
