% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{is.RI.regression.result}
\alias{is.RI.regression.result}
\title{#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'}
\usage{
is.RI.regression.result(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Clean printing of our result objects
}
\details{
#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'#'
Check if object is a RI.regression.result class
}
