% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{plot.RI.R2.result}
\alias{plot.RI.R2.result}
\title{Make a plot of the treatment effect R2 estimates}
\usage{
\method{plot}{RI.R2.result}(x, main = paste("R2 for Het Tx (", x$type,
  ")", sep = ""), ADD = FALSE, ...)
}
\arguments{
\item{x}{Results from est.beta, etc.}

\item{main}{Title for plot}

\item{ADD}{TRUE if add to existing plot. FALSE make a new plot.}

\item{...}{Arguments to pass to plotting of points.}
}
\description{
Make a plot of the treatment effect R2 estimates
}
\seealso{
calc.beta
}
