% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{print.RI.regression.result}
\alias{print.RI.regression.result}
\title{Print out result of systematic treatment effect estimation.}
\usage{
\method{print}{RI.regression.result}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{Result from est.beta}

\item{digits}{Number of digits for rounding}

\item{...}{Unused}
}
\description{
Print out result of systematic treatment effect estimation.
}
