% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{R2.ITT}
\alias{R2.ITT}
\title{using the OLS output to bound the R2 measure}
\usage{
R2.ITT(RI.result, rho.step = 0.05)
}
\arguments{
\item{RI.result}{RI regression result (from est.beta, e.g.,) to process.}

\item{rho.step}{Grid size for sensitivity analysis on values of rho.}
}
\value{
RI.R2.result object.
}
\description{
using the OLS output to bound the R2 measure
}
\seealso{
print.RI.R2.result
systematic component
idiosyncratic component
approximate the intergral by summation--step quantile functions
R2: bounds and sensitivity analysis
results
}
