% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{R2.LATE}
\alias{R2.LATE}
\title{using the LATE estimation output to bound the R2 measures for compliers, etc.}
\usage{
R2.LATE(RI.result, rho.step = 0.05)
}
\arguments{
\item{RI.result}{RI regression result (from est.beta, e.g.,) to process.}

\item{rho.step}{Grid size for sensitivity analysis on values of rho.}
}
\description{
using the LATE estimation output to bound the R2 measures for compliers, etc.
}
\seealso{
R2.ITT
tau.c: Wald estimator (LATE estimate)
systematic component by U
systematic component by X for compliers
evaluate the ECDFs at the unique values of residuals = Yunique
approximate the intergral by summation--step quantile functions
initial values of the integrands
three R2: bounds and sensitivity analysis
R2.U
R2.tau.c
R2.UX
results
}
