% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_calc_helpers.R
\name{SKS.stat.cov.pool}
\alias{SKS.stat.cov.pool}
\title{SKS.stat.cov.pool}
\usage{
SKS.stat.cov.pool(Y, Z, X)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}

\item{X}{Additional pre-treatment covariates to adjust for in estimation, but not to interact with treatment. Default is NULL.}
}
\value{
The value of the test.
}
\description{
Shifted kolmogorov-smirnov statistic with covariates
to increase precision.
}
\details{
This is the test statistic used Ding, Feller, and Miratrix (2016), JRSS-B.
}
