% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_calc_helpers.R
\name{SKS.stat.cov}
\alias{SKS.stat.cov}
\title{SKS.stat.cov}
\usage{
SKS.stat.cov(Y, Z, X)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}

\item{X}{Additional pre-treatment covariates to adjust for in estimation, but not to interact with treatment. Default is NULL.}
}
\value{
The value of the test.
}
\description{
Shifted kolmogorov-smirnov statistic with covariates
with model for outcomes calculated on control group only.
This avoids "splitting" the treatment variation between tx 
and co groups.  
We recommend this method over the "pool" method.
}
