% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToyData}
\alias{ToyData}
\title{Toy data set}
\format{A dataframe containing 500 observations and 7 columns.}
\usage{
ToyData
}
\description{
This is a toy data set to illustrate the package methods.
}
\keyword{dataset}
