% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{variance.ratio.test}
\alias{variance.ratio.test}
\title{Variance ratio test}
\usage{
variance.ratio.test(Yobs, Z, data = NULL)
}
\arguments{
\item{Yobs}{Outcome}

\item{Z}{Treatment assignment vector}

\item{data}{Dataframe with variables listed in formula and control.formula}
}
\description{
Given vector of observed outcomes and treatment vector, test to see if there is evidence
the variances are different (taking kurtosis into account).
}
