% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{calc.beta.oracle}
\alias{calc.beta.oracle}
\title{Calculate systematic effects model with full potential outcomes}
\usage{
calc.beta.oracle(formula, data, interaction.formula, method = c("RI",
  "OLS"), na.rm = FALSE)
}
\arguments{
\item{formula}{An object of class formula, as in lm(), such as Y(1) + Y(0) ~ Z with only the treatment variable on the right-hand side and the variables indicating the outcome under treatment and the outcome under control on the left-hand-side. The first variable on the left-hand-side will be treated as the outcome under treatment, and the second variable on the right-hand-side will be treated as the outcome under control.}

\item{data}{A data.frame, tbl_df, or data.table with the input data.}

\item{interaction.formula}{A right-sided formula with pre-treatment covariates to model treatment effects for on the right hand side, such as ~ x1 + x2 + x3.}

\item{method}{RI or OLS.}

\item{na.rm}{A logical flag indicating whether to list-wise delete missing data. The function will report an error if missing data exist. Default is FALSE.}
}
\description{
Calculate the actual beta for our sample using the full
potential outcome schedule (if available)
}
\details{
This function is useful for sanity checks and simulation studies
to compare real estimators to optimal baselines.

Not useful for observed data as it relys on complete knowledge of all potential outcomes.
}
\seealso{
make.randomized.dat
covariance between X and tau
covariance between X and Y(0)
covariance between X and Y(1)
Calculate individual systematic effects
return results
}
