% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{est.beta.LATE}
\alias{est.beta.LATE}
\title{Estimate systematic effects in presence of noncompliance}
\usage{
est.beta.LATE(Z, D, Yobs, formula, data, method = c("RI", "2SLS"))
}
\arguments{
\item{Z}{Treatment assignment vector}

\item{D}{treatment received}

\item{Yobs}{Outcome}

\item{formula}{Formula describing what variables to model treatment effect for.}

\item{data}{Dataframe with variables listed in formula and control.formula}

\item{method}{RI or OLS.  method=OLS is shorthand for setting the empirical.Sxx variable to TRUE, nothing more.}
}
\description{
Estimate a model of effects for compliers under exclusion restriction for
never- and always-takers and under monotonicity.
}
\seealso{
est.beta
sample size, and sub-sample sizes classified by (Z, D)
estimation of the proportions of the latent strata
TSLS
residual
gamma for compliers: under treatment
gamma for compliers: under control
unbiased estimator for beta
residual
covariance of beta
testing
return
}
