% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{est.beta}
\alias{est.beta}
\title{Estimate Systematic Treatment Effect Variation}
\usage{
est.beta(Yobs, Z, formula, control.formula = NULL, data, method = c("RI",
  "OLS"), empirical.Sxx = FALSE)
}
\arguments{
\item{Yobs}{Outcome}

\item{Z}{Treatment assignment vector}

\item{formula}{Formula describing what variables to model treatment effect for.}

\item{control.formula}{Formula for what variables to adjust for.  If we have a control formula, we adjust}

\item{data}{Dataframe with variables listed in formula and control.formula}

\item{method}{RI or OLS.  method=OLS is shorthand for setting the empirical.Sxx variable to TRUE, nothing more.}

\item{empirical.Sxx}{Estimate seperate Sxx for treatment and control if TRUE, use known Sxx if not.}
}
\description{
Estimate our treatment effect model given by the formula "~ X"
Will do the different estimators listed in our paper given the 'method' flag
}
\seealso{
print.RI.regression.result
sample size
dimension of X
treatment group
control group
calculate covariance matrix of X and estimate our gammas
residuals
estimate beta
final estimator for tau, the individual systematic treatments
covariance
testing whether any of the non-intercept terms are nonzero
return results
}
